backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
reportlab==4.0.7
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
numpy>=1.26.0,<2.0.0